CAMEL-22894: Split SimpleExpressionBuilder into domain-aligned builder classes
Replace the monolithic SimpleExpressionBuilder (3,523 lines, 100+ static
methods) with seven focused builder classes:

- CollectionExpressionBuilder: list, map, filter, sort, set-header, etc.
- DateExpressionBuilder:       date/time expressions with offset support
- FileExpressionBuilder:       file name, path, size, extension, etc.
- MathExpressionBuilder:       abs, floor, ceil, sum, min, max, average
- MiscExpressionBuilder:       uuid, random, type, hash, convert, etc.
- OgnlExpressionBuilder:       OGNL body/header/variable/exchange navigation
- StringExpressionBuilder:     pad, trim, substring, replace, contains, etc.

KeyedOgnlExpressionAdapter (inner class) moves to OgnlExpressionBuilder.
OFFSET_PATTERN static field moves to DateExpressionBuilder.

All 22 function factories and ast/SimpleFunctionExpression updated to import
their respective domain builder. AttachmentExpressionBuilder (camel-attachments)
updated to use OgnlExpressionBuilder.KeyedOgnlExpressionAdapter.

Pure mechanical refactor: no behavioral change.


rh-pre-commit.version: 2.4.0
rh-pre-commit.check-secrets: ENABLED
