+++ core/camel-core-languages/src/main/java/org/apache/camel/language/simple/functions/CollateFunctionFactory.java
@@ -18,7 +18,7 @@ package org.apache.camel.language.simple.functions;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Expression;
-import org.apache.camel.language.simple.SimpleExpressionBuilder;
+import org.apache.camel.language.simple.MiscExpressionBuilder;
 import org.apache.camel.language.simple.types.SimpleParserException;
 import org.apache.camel.spi.SimpleLanguageFunctionFactory;
 import org.apache.camel.util.ObjectHelper;
@@ -41,7 +41,7 @@ public final class CollateFunctionFactory implements SimpleLanguageFunctionFacto
         if (values == null || ObjectHelper.isEmpty(values)) {
             throw new SimpleParserException("Valid syntax: ${collate(group)} was: " + function, index);
         }
-        return SimpleExpressionBuilder.collateExpression("${body}", values.trim());
+        return MiscExpressionBuilder.collateExpression("${body}", values.trim());
     }
 
     @Override
