+++ core/camel-core-languages/src/main/java/org/apache/camel/language/simple/functions/CustomFunctionFactory.java
@@ -18,7 +18,7 @@ package org.apache.camel.language.simple.functions;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Expression;
-import org.apache.camel.language.simple.SimpleExpressionBuilder;
+import org.apache.camel.language.simple.MiscExpressionBuilder;
 import org.apache.camel.language.simple.types.SimpleParserException;
 import org.apache.camel.spi.SimpleLanguageFunctionFactory;
 import org.apache.camel.util.ObjectHelper;
@@ -61,7 +61,7 @@ public final class CustomFunctionFactory implements SimpleLanguageFunctionFactor
         if (param == null) {
             param = "${body}";
         }
-        return SimpleExpressionBuilder.customFunction(key, param);
+        return MiscExpressionBuilder.customFunction(key, param);
     }
 
     @Override
